First check-in of Skein functions for BouncyCastle environment. This Skein implementation is based on Alberto Fajardo's C# implementation. I edited the Java version and did some enhancements to have full unrolled loops without functions call as done in C# implementation. Several tests with test vectors were successful.
All three Skein state sizes (24´56, 512, and 1024) are supported, flexible
hash output sizes are supported, and as a special bonus: you can hash messages
up to a length of 96 bits (maximum message length for Skein).

Based the the Skein I added functions and classes to use Skein as MAC.

The check-in also has a Threefish encrypt / decrypt which is not yet
integrated into BouncyCastle.