Modify ccrtp/crypto sources to match the master sources.
The sources in ccrtp/crypto are now in sync with the master sources in libzrtpcpp/srtp/crypto (see
my libzrtpcpp project at GitHub). The modifications als required some small changes in the build
files (cmake and automake). Currently only the realy SRTP crypto contexts are not yet in sync
and must be handled separately in necessary.